Add stopping point
