--- Assets/ScriptableObjects/Biomes.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2551157429d9448b595eb0ba93c8d0b8
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
