--- Assets/ScriptableObjects/Grid Manager.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 08a6b1332d45c4c8db970ca0669179d9
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
