+++ Assets/ScriptableObjects/World.asset
@@ -10,7 +10,7 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 6210209ff49ad455c931ff2e0d6fa996, type: 3}
-  m_Name: All Biomes
+  m_Name: World
   m_EditorClassIdentifier: 
   HeightMin: 0.09847487
   HeightMax: 0.8426731
