+++ Assets/Scripts/_Managers/GridManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 47ccb23bbd6104f0ebe01ba8d01a7808
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
