+++ Assets/Scripts/_Unused/Tiles/Copper Ore.mat
